Some fixes
