+++ package.json
@@ -35,7 +35,7 @@
     "clsx": "^1.1.1",
     "dayjs": "^1.9.6",
     "formik": "^2.2.6",
-    "formik-mui": "*",
+    "formik-mui": "^4.0.0-alpha.3",
     "lodash": "^4.17.15",
     "qs": "^6.9.1",
     "react": "^17.0.2",
