+++ src/apps/authentication/components/UserForm.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef } from "react";
 import { makeStyles } from '@mui/styles';
 import { Field, Form, Formik, FormikProps, FormikValues } from "formik";
-import { TextField } from "formik-material-ui";
+import { TextField } from "formik-mui";
 import { APIProfile } from "../types";
 import dayjs from "dayjs";
 import { omitBy, isNull } from "lodash";
