+++ src/index.tsx
@@ -1,32 +1,12 @@
-import React from "react";
 import ReactDOM from "react-dom";
-import { ThemeProvider, createTheme } from "@mui/material/styles";
 import "./fonts/nunito.css";
 import "./index.scss";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
 
-export const theme = createTheme({
-  palette: {
-    primary: {
-      light: "#757ce8",
-      main: "#3f51b5",
-      dark: "#002884",
-      contrastText: "#fff"
-    },
-    secondary: {
-      light: "#ff7961",
-      main: "#f50057",
-      dark: "#ba000d",
-      contrastText: "#000"
-    }
-  }
-});
 
 ReactDOM.render(
-  <ThemeProvider theme={theme}>
-    <App />
-  </ThemeProvider>,
+  <App />,
   document.getElementById("root")
 );
